Tableau Map Visualizations - Non Residential Meter Readings Ireland and Leinster Area
Three clusters have been set for each visualization, where are identified the counties in Ireland with the large consumption in electricity in Ireland. Similarly, the main consumers of Electricity in Leinster area have been shown as well. Data Centres are located in the counties with the large electricity consumption in Ireland.